Move meta API group out of internal.
